# ts-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "fulcrumaxe ts-backend API",
    "version": "0.1.0",
    "description": "TypeScript/Bun+Hono backend — mirrors the Python backend with 1:1 parity. Runs on 127.0.0.1:19099 (loopback-only). Auth: set AF_API_AUTH_KEY to enable; unset = auth disabled. SSE routes (/feed, /events) accept ?token= for EventSource clients that cannot send custom headers. See https://github.com/fulcrumaxe/fulcrumaxe/discussions/1437 for the full phased migration plan.",
    "contact": {
      "url": "https://github.com/fulcrumaxe/fulcrumaxe"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:19099",
      "description": "Default loopback server (TS_BACKEND_PORT overrides 19099)"
    }
  ],
  "security": [],
...
  "paths": {
...
    "/feed": {
      "get": {
        "summary": "Agent feed SSE stream",
        "description": "Server-Sent Events stream that tails .autonomous-team/agent-feed.jsonl. Auth: Bearer header OR ?token= query param (EventSource-compatible). Initial frame: {type:\"connected\"}. Heartbeat every 30s idle: {type:\"heartbeat\"}. Poll interval: 500ms. Supports ?since=<iso> and ?filter[role]=<role> filters.",
        "operationId": "getFeedStream",
        "tags": [
          "Streaming"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Bearer token as a query param. Use when Authorization header is not settable (EventSource clients)."
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "ISO 8601 timestamp. Only emit events with ts >= this value."
          },
          {
            "name": "filter[role]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter to events where role matches this value exactly."
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of agent feed events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SseFrame"
                }
              }
            },
            "headers": {
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "text/event-stream; charset=utf-8"
                }
              },
              "X-Accel-Buffering": {
                "schema": {
                  "type": "string",
                  "example": "no"
                },
                "description": "Disables nginx buffering for real-time SSE delivery"
              }
            }
          },
          "401": {
            "description": "No credentials provided (auth enabled, no token)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailError"
                }
              }
            }
          }
        }
      }
    },
    "/events": {
      "get": {
        "summary": "Events SSE stream (typed)",
        "description": "Server-Sent Events stream that tails agent-feed.jsonl and adds an _event_type discriminator field to each event. Event types inferred: BudgetSpendEvent, GateChangeEvent, LoopIterationEvent, AgentOutputEvent. Auth: Bearer header OR ?token= query param. Supports ?loop_id= and ?since= filters.",
        "operationId": "getEventsStream",
        "tags": [
          "Streaming"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Bearer token as query param (EventSource-compatible auth)."
          },
          {
            "name": "loop_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter to events from a specific loop run."
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "ISO 8601 timestamp. Only emit events with ts >= this value."
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of typed events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SseFrame"
                }
              }
            },
            "headers": {
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "text/event-stream; charset=utf-8"
                }
              },
              "X-Accel-Buffering": {
                "schema": {
                  "type": "string",
                  "example": "no"
                }
              }
            }
          },
          "401": {
            "description": "No credentials provided (auth enabled, no token)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token from AF_API_AUTH_KEY. Required when auth is enabled. SSE routes (/feed, /events) also accept ?token=<key> as a query param because EventSource clients cannot set custom headers."
      }
    },
    "schemas": {
      "DetailError": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable error message"
          }
        }
      },
...
      "SseFrame": {
        "type": "string",
        "description": "Server-Sent Events wire format: each event is \"data: <json>\\n\\n\". The initial frame is {type:\"connected\"}, heartbeats are {type:\"heartbeat\"} (every 30s idle). Actual event payloads are JSONL records from agent-feed.jsonl.",
        "example": "data: {\"type\":\"connected\"}\n\ndata: {\"event_type\":\"agent_run\",...}\n\n"
      }
    }
  }
}